requirement.txt
Flask
pandas
scikit-learn
SQLAlchemy
psycopg2-binary
python-dotenv
numpy
Werkzeug